Move colorization logic into respective components
